Add pixel coordinates, mouseover function
The character coordinates were initially in terms of the tile map, which did not
correspond directly to the coordinates of the position on the screen. Entities
now have a xPixel and yPixel attribute that records their position on the screen
itself. From this, a function mouseoverQ, which is a member of the Entities
class will be able to tell when a entity is being hovered over with a mouse.

Note, that updating the pixel coordinates is done within the update function,
which now takes the camera as an argument. As the child objects of entities
overrides the update function, the actual update must take place there, e.g. in
the update function of the Character child class.