Fix off-center sprite renering
There was a scaling issue where the character sprite would be rendering in the
correct position when it was original size, but after scaling it would be off
the grid. The resolution was the set the origin of the sprite to (0,0), which is
top left. (This doesn't make complete sense as it says the default should be the
bottom left, but we're going to go with it.)